INS-213: switch all ledger tests on one store implementation (BadgerDB by default)
